✨ feat: 상품 CURD 수정 및 로직 추가
- Products price,stock 데이터타입 변경, 재고상태값 변경 로직 추가, setter 제거
- ProductRequest price,stock 데이터타입 변경
- ProductResponse price,stock 데이터타입 변경, ProductStatus 추가
- ProductService 저장 로직을 검증 이후로 변경

+++ src/main/java/com/example/lineofduty/domain/product/dto/request/ProductRequest.java
@@ -1,6 +1,5 @@
 package com.example.lineofduty.domain.product.dto.request;
 
-import com.example.lineofduty.common.exception.ValidationMessage;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -17,9 +16,9 @@ public class ProductRequest {
 
     @NotNull
     @Min(value = 0)
-    private Integer price;
+    private Long price;
 
     @NotNull
     @Min(value = 0)
-    private Integer stock;
+    private Long stock;
 }
